Main change: Add 'enable_sshd' and disable_winrm
Also: Some refactoring and improvement of argument
handling in files/vm.boxstarter.  And some
cleanups with the way that options are handled.

+++ scripts/disable-winrm-and-shutdown.ps1
@@ -1,12 +1,21 @@
 # Uncomment for debugging
 #Start-Transcript C:/debug.log -Append
 
+Param(
+    [Switch]$DisableWinRM
+)
+
 Start-Sleep -Seconds 15
 
-Write-Host "Disabling WinRM"
-Disable-PSRemoting -Force
-Set-Service WinRM -StartupType Disabled -PassThru
-Remove-Item -Path WSMan:\Localhost\listener\listener* -Recurse
-Set-NetFirewallRule -DisplayName 'Windows Remote Management (HTTP-In)' -Enabled False -PassThru | Select -Property DisplayName, Profile, Enabled
+if ($DisableWinRM) {
+    Write-Host "Disabling WinRM"
+    Disable-PSRemoting -Force
+    Set-Service WinRM -StartupType Disabled -PassThru
+    Remove-Item -Path WSMan:\Localhost\listener\listener* -Recurse
+    Set-NetFirewallRule -DisplayName 'Windows Remote Management (HTTP-In)' -Enabled False -PassThru | Select -Property DisplayName, Profile, Enabled
+} else {
+    Write-Host "Skipping WinRM disable (DisableWinRM = $DisableWinRM)"
+}
+
 Write-Host "Shutting down now"
 shutdown /s /f /t 0 /c "Packer has finished. Shutting down."
